+++ Pod/SyneriseSDK.framework/Headers/SNRInvalidBirthdateError.h
@@ -1,0 +1,11 @@
+// 
+//  SNRInvalidBirthdateError.h
+//  SyneriseSDK
+//
+//  Created by Synerise
+//  Copyright (c) 2018 Synerise. All rights reserved.
+//
+
+@interface SNRInvalidBirthdateError : SNRError
+
+@end