+++ Pod/SyneriseSDK.framework/Headers/SNRInvalidBirthdateException.h
@@ -1,0 +1,11 @@
+// 
+//  SNRInvalidBirthdateException.h
+//  SyneriseSDK
+//
+//  Created by Synerise
+//  Copyright (c) 2018 Synerise. All rights reserved.
+//
+
+@interface SNRInvalidBirthdateException : SNRException
+
+@end