+++ Pod/SyneriseSDK.framework/Headers/SNRInvalidPasswordError.h
@@ -1,0 +1,11 @@
+// 
+//  SNRInvalidPasswordError.h
+//  SyneriseSDK
+//
+//  Created by Synerise
+//  Copyright (c) 2018 Synerise. All rights reserved.
+//
+
+@interface SNRInvalidPasswordError : SNRError
+
+@end