+++ Pod/SyneriseSDK.framework/Headers/SNRInvalidPasswordException.h
@@ -1,0 +1,11 @@
+// 
+//  SNRInvalidPasswordException.h
+//  SyneriseSDK
+//
+//  Created by Synerise
+//  Copyright (c) 2018 Synerise. All rights reserved.
+//
+
+@interface SNRInvalidPasswordException : SNRException
+
+@end