+++ CHANGELOG.md
@@ -1,10 +1,21 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [3.2.12] - 2018-06-22
+
+### Added
+- Method to get current client's UUID.
+- customId property in ClientProfileContext.
+- improvements for exceptions/errors handling in obj-c/swift.
+
+### Fixed
+- Profile.createClient.
+
+
 ## [3.2.11] - 2018-06-21
 
 ### Fixed
-- Profile.deleteClient
+- Profile.deleteClient.
 
 
 ## [3.2.10] - 2018-06-18